+++ src/Compilers/Core/Portable/SourceGeneration/Nodes/SyntaxReceiverStrategy.cs
